% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claimsBoot-package.R
\docType{package}
\name{claimsBoot}
\alias{claimsBoot}
\alias{claimsBoot-package}
\title{claimsBoot}
\description{
Description of your package
}
\author{
Othman El Hammouchi <othman.el.hammouchi@hotmail.com>
}
\keyword{internal}
