% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cum2incr}
\alias{cum2incr}
\title{Convert cumulative triangle to incremental one.}
\usage{
cum2incr(triangle.cum)
}
\arguments{
\item{triangle.cum}{Cumulative claims triangle.}
}
\value{
Incremental claims triangle.
}
\description{
Convert cumulative triangle to incremental one.
}
