% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glmBoot}
\alias{glmBoot}
\title{Simulate Poisson GLM reserve.}
\usage{
glmBoot(triangle, n_boot, boot_type, opt = "null", seed = 42L)
}
\arguments{
\item{triangle}{Incremental claims triangle}
}
\description{
Simulate Poisson GLM reserve.
}
