% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.r
\name{glmSemiParamBoot}
\alias{glmSemiParamBoot}
\title{glmSemiParamBoot}
\usage{
glmSemiParamBoot(triangle, nboot, nsim, resids.type)
}
\arguments{
\item{triangle}{Incremental claims triangle.}

\item{nboot}{Number of bootstrap iterations.}

\item{nsim}{Number of simulation iterations.}

\item{resids.type}{The type of residuals to use: `"pearson"` or `"deviance"`.}
}
\value{
A list with the following components:
* A dataframe `table` containing the estimates for the a and b parameters as well as
  the reserve by origin year. Estimates of the prediction error by origin year are provided
  as well.
* A dataframe `point` containing point estimates for the intercept and dispersion.
* A vector `pred.dist` containing the simulated predictive distribution of the reserve.
}
\description{
Perform a semiparametric bootstrap for the GLM reserving model.
}
