% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{incr2cum}
\alias{incr2cum}
\title{Convert incremental triangle to cumulative one.}
\usage{
incr2cum(triangle.incr)
}
\arguments{
\item{triangle.inrc}{Incremental claims triangle.}
}
\value{
Cumulative claims triangle.
}
\description{
Convert incremental triangle to cumulative one.
}
