% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mackBoot}
\alias{mackBoot}
\title{Simulate Mack CL reserve.}
\usage{
mackBoot(triangle, n_boot, boot_type, opt, cond, seed = 42L)
}
\arguments{
\item{triangle}{Cumulative claims triangle}

\item{n_boot}{Number of bootstrap simulations}

\item{boot_type}{Type of bootstrap: `"parametric"`, `"residuals"`, or `"pairs"`}

\item{cond}{Specified whether the bootstrap should be conditional or unconditional. Default is `TRUE`.}

\item{proc_dist}{Distribution of process error: `"normal"`or `"gamma"`}
}
\description{
Simulate Mack CL reserve.
}
