% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack.r
\name{mackPairsBoot}
\alias{mackPairsBoot}
\title{Perform pairs bootstrap for the Mack chain ladder}
\usage{
mackPairsBoot(triangle, nboot, nsim, sim.type)
}
\arguments{
\item{triangle}{Cumulative claims triangle.}

\item{nboot}{Number of bootstrap iterations.}

\item{nsim}{Number of iterations for the predictive distribution.}

\item{sim.type}{Type of simulation for the predictive distribution.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
