% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack.r
\name{mackParamBoot}
\alias{mackParamBoot}
\title{FUNCTION_TITLE}
\usage{
mackParamBoot(triangle, nboot, nsim, cond, dist)
}
\arguments{
\item{triangle}{DESCRIPTION.}

\item{nboot}{DESCRIPTION.}

\item{nsim}{DESCRIPTION.}

\item{cond}{DESCRIPTION.}

\item{dist}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
