% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack.r
\name{mackSemiParamBoot}
\alias{mackSemiParamBoot}
\title{Perform semiparametric bootstrap for the Mack chain ladder}
\usage{
mackSemiParamBoot(triangle, nboot, nsim, cond, resids.type)
}
\arguments{
\item{triangle}{DESCRIPTION.}

\item{nboot}{DESCRIPTION.}

\item{nsim}{DESCRIPTION.}

\item{cond}{DESCRIPTION.}

\item{resids.type}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
