% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.r
\name{glmPairsBoot}
\alias{glmPairsBoot}
\title{FUNCTION_TITLE}
\usage{
glmPairsBoot(triangle, nboot, nsim, keep)
}
\arguments{
\item{triangle}{Incremental claims triangle.}

\item{nboot}{DESCRIPTION.}

\item{nsim}{DESCRIPTION.}

\item{keep}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
