% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.r
\name{glmParamBoot}
\alias{glmParamBoot}
\title{FUNCTION_TITLE}
\usage{
glmParamBoot(triangle, nboot, nsim, dist)
}
\arguments{
\item{triangle}{DESCRIPTION.}

\item{nboot}{DESCRIPTION.}

\item{nsim}{DESCRIPTION.}

\item{dist}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
