% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.r
\name{glmSemiParamBoot}
\alias{glmSemiParamBoot}
\title{FUNCTION_TITLE}
\usage{
glmSemiParamBoot(triangle, nboot, nsim, resids.type)
}
\arguments{
\item{triangle}{DESCRIPTION.}

\item{nboot}{DESCRIPTION.}

\item{nsim}{DESCRIPTION.}

\item{resids.type}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
